mycode/image-2-pdf-ocrizer/requirements.txt
chardet==5.2.0
fpdf==1.7.2
MouseInfo==0.1.3
numpy==2.2.1
opencv-python-headless==4.10.0.84
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyocr==0.8.5
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python3-xlib==0.15
pytweening==1.2.0
reportlab==4.2.5
